+++ tests/accounts/manage.py
@@ -1,6 +1,6 @@
 from nacl.signing import SigningKey, VerifyKey
 
-from tnb.accounts.accounts import create_account
+from tnb.accounts.manage import create_account
 
 
 def test_create_account():
